+++ playground/control/attribs/__init__.py
@@ -1,0 +1,19 @@
+
+from .base import (
+    AttribsWithName, ValidatingAttribs)
+from .request import (
+    NetworkAddAttribs, NetworkDeleteAttribs, NetworkEditAttribs,
+    ProxyAddAttribs, ProxyDeleteAttribs,
+    ServiceAddAttribs, ServiceDeleteAttribs)
+
+
+__all__ = (
+    'AttribsWithName',
+    'NetworkAddAttribs',
+    'NetworkDeleteAttribs',
+    'NetworkEditAttribs',
+    'ProxyAddAttribs',
+    'ProxyDeleteAttribs',
+    'ServiceAddAttribs',
+    'ServiceDeleteAttribs',
+    'ValidatingAttribs')
