+++ playground/control/attribs/request/__init__.py
@@ -1,0 +1,17 @@
+
+from .network import (
+    NetworkAddAttribs, NetworkDeleteAttribs, NetworkEditAttribs)
+from .proxy import (
+    ProxyAddAttribs, ProxyDeleteAttribs)
+from .service import (
+    ServiceAddAttribs, ServiceDeleteAttribs)
+
+
+__all__ = (
+    'NetworkAddAttribs',
+    'NetworkDeleteAttribs',
+    'NetworkEditAttribs',
+    'ProxyAddAttribs',
+    'ProxyDeleteAttribs',
+    'ServiceAddAttribs',
+    'ServiceDeleteAttribs')
