+++ tests/py/validators.py
@@ -1,6 +1,6 @@
 
-import playground.control.validators
+import playground.control.attribs.validators
 
 
 def test_validators():
-    assert playground.control.validators is not None
+    assert playground.control.attribs.validators is not None
